Updates imports according to renamed components
